Refactored lib0.hpp - separated out+removed old; [doc] ornament
