+++ src/libs0.cpp
@@ -1,25 +1,7 @@
 // Copyrighted (C) 2015-2016 Antinet.org team, see file LICENCE-by-Antinet.txt
 
-
 #include "libs0.hpp"
 
-
-const char* expected_exception::what() const noexcept {
-	return "expected_exception";
-}
-
-const char* expected_not_found::what() const noexcept {
-	return "expected_not_found";
-}
-
-
-namespace stdplus {
-
-std::string to_string(const std::string & v) {
-	return v;
-}
-
-} // namespace
 
 #if defined(_MSC_VER)
 const char * gettext(const char * s) {
