+++ src/stdplus/tab.hpp
@@ -1,0 +1,22 @@
+// Copyrighted (C) 2015-2017 Antinet.org team, see file LICENCE-by-Antinet.txt
+
+#pragma once
+
+/**
+@file stdplus is library of very common functions to be used in this and other projects,
+stdplus/tab focuses on containers support
+@owner rfree
+@todo TODO@testcase - test cases for all functions here
+*/
+
+
+namespace stdplus {
+
+
+
+
+
+} // namespace stdplus
+
+
+
